+++ source/Properties.pas
@@ -10,44 +10,68 @@ Type
   private
     foProject: tProject;
     FoOutput: TOutput;
+    foProjectItem: tObject;
   public
+    constructor Create(aProjectItem: tObject);  reintroduce;
+
     function IsPropertyExists(APropertyName: String): Boolean;
     function GetProperty(APropertyName: String): String;
-    function DoPluginProperty(APropertyName: string): string;
+    //function GetPluginProperty(APropertyName: string): string;
 
-    property oProject: tProject
-      read foProject
-      write foProject;
+//    property oProject: tProject
+//      read foProject
+//      write foProject;
 
-    property oOutput: TOutput
-      read FoOutput
-      write FoOutput;
+//    property oOutput: TOutput
+//      read FoOutput
+//      write FoOutput;
   end;
 
 implementation
 
-uses Runtime, Plugin, ProjectParser;
+uses Runtime, Plugin, ProjectParser, variables, ProjectItem;
+
+constructor tProperties.Create(aProjectItem: tObject); //virtual;
+begin
+  inherited Create;
+
+  foProjectItem := aProjectItem;
+
+  if Assigned(foProjectItem) then
+    begin
+      FoOutput := (foProjectItem as TProjectItem).oOutput;
+      foProject := (foProjectItem as TProjectItem).oPRoject;
+    end;
+end;
 
 
 function tProperties.GetProperty(APropertyName: String): String;
 var
   lsGetProperty: String;
+  fVariable: tVariable;
 begin
+
   lsGetProperty := GetFieldAsString(oXMLDocument.Root, APropertyName);
   Try
     if Assigned(foProject) then
       if foProject.oProjectConfigLoader.Load then
          lsGetProperty := tProjectParser.ParseProject(lsGetProperty,foProject, foOutput);
+
+    fVariable := (foProjectItem as TProjectItem).oVariables.GetVariableByName(APropertyName);
+    if assigned(fVariable) then
+      lsGetProperty := fVariable.AsString;
   Except
     FoOutput.log(APropertyName + ' Projectconfig error.');
 
   End;
 
   Result := lsGetProperty;
+
 end;
 
 
-function tProperties.DoPluginProperty(APropertyName: string): string;
+(*
+function tProperties.GetPluginProperty(APropertyName: string): string;
 var
   loPlugin: TPlugin;
   I: Integer;
@@ -58,6 +82,8 @@ begin
 
     end;
 end;
+*)
+
 
 function tProperties.IsPropertyExists(APropertyName: String): Boolean;
 Var